+++ assets/scripts/app.js
@@ -22,6 +22,7 @@ ipcRenderer.on('plot-timeseries', function(event, data) {
 d3.select('#new-session').on('click', () => caim.new_session());
 d3.select('#open-session').on('click', () => caim.open_session());
 d3.select('#import-video').on('click', () => caim.import_video());
+d3.select('#import-frames').on('click', () => caim.import_frames());
 
 d3.select('#clear').on('click', () => caim.clear());
 d3.select('#undo').on('click', () => caim.undo());