+++ CHANGELOG.md
@@ -5,13 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+* Import frames from images in a directory ([#20](issues/20))
+
 ## [0.3.0] - 2020-02-08
 
 ### Added
 
 * Grid selection
 * Mask selection
-* Fix FeatureGroup loading on Session.load
+
+### Fxed
+* FeatureGroup loading on Session.load
 
 ## [0.2.0] - 2019-09-23